Forms: Rename icon.js files to icon.jsx for proper JSX extensions
