+++ projects/packages/forms/src/blocks/field-date/index.js
@@ -1,6 +1,6 @@
 import { Path } from '@wordpress/components';
 import { __, _x } from '@wordpress/i18n';
-import renderMaterialIcon from '../shared/components/render-material-icon.js';
+import renderMaterialIcon from '../shared/components/render-material-icon.jsx';
 import defaultSettings from '../shared/settings/index.js';
 import deprecated from './deprecated.js';
 import edit from './edit.js';
