+++ projects/packages/forms/src/blocks/field-email/icon.jsx
