+++ projects/packages/forms/src/blocks/field-rating/variations.js
@@ -1,6 +1,6 @@
 import { Path } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import renderMaterialIcon from '../shared/components/render-material-icon.js';
+import renderMaterialIcon from '../shared/components/render-material-icon.jsx';
 import { RATING_ICONS } from './rating-icons.js';
 
 /**
