+++ projects/packages/forms/src/blocks/field-select/icon.jsx
@@ -1,5 +1,5 @@
 import { Path } from '@wordpress/components';
-import renderMaterialIcon from '../shared/components/render-material-icon.js';
+import renderMaterialIcon from '../shared/components/render-material-icon.jsx';
 
 export default {
 	src: renderMaterialIcon(
