--- projects/packages/forms/src/blocks/field-slider/icon.js
@@ -1,5 +1,0 @@
-import SliderIcon from '../input-range/icon.js';
-
-export default {
-	src: <SliderIcon />,
-};
