+++ projects/packages/forms/src/blocks/field-slider/icon.jsx
@@ -1,0 +1,5 @@
+import SliderIcon from '../input-range/icon.jsx';
+
+export default {
+	src: <SliderIcon />,
+};
