+++ projects/packages/forms/src/blocks/field-telephone/icon.jsx
