+++ projects/packages/forms/src/blocks/input-image-option/icon.tsx
@@ -5,7 +5,7 @@ import { Path } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import renderMaterialIcon from '../shared/components/render-material-icon.js';
+import renderMaterialIcon from '../shared/components/render-material-icon.jsx';
 
 // "group" icon from @wordpress/icons
 const ImageOptionInputIcon = renderMaterialIcon(
