+++ projects/packages/forms/src/blocks/input-range/icon.jsx
